api_project/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-authtoken